+++ workers/unity/Packages/io.improbable.gdk.deploymentlauncher/Tests/Editmode/DeploymentConfig/Utils/DeploymentConfigErrorTypes.cs
@@ -1,0 +1,11 @@
+namespace Improbable.Gdk.DeploymentLauncher.EditmodeTests.Utils
+{
+    internal static class DeploymentConfigErrorTypes
+    {
+        internal const string Invalid = "invalid";
+        internal const string NullOrEmpty = "empty";
+        internal const string FullStops = "full stops";
+        internal const string Spaces = "spaces";
+        internal const string NotFound = "cannot be found";
+    }
+}